Move LdtkSpriteSheetBundle it its own file
