+++ Cargo.toml
@@ -21,7 +21,6 @@ derive-getters = "0.3.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 regex = "1"
-hex = "0.4"
 thiserror = "1.0"
 paste = "1.0"
 derive_more = "0.99.17"
