+++ src/app/ldtk_entity.rs
@@ -1,7 +1,6 @@
 use crate::{
     components::{EntityInstanceBundle, GridCoords, Worldly},
     ldtk::{EntityInstance, LayerInstance, TilesetDefinition},
-    prelude::LdtkSpriteSheetBundle,
     utils,
 };
 use bevy::{ecs::system::EntityCommands, prelude::*};
@@ -358,25 +357,6 @@ impl LdtkEntity for SpriteBundle {
         _: &mut Assets<TextureAtlasLayout>,
     ) -> Self {
         utils::sprite_bundle_from_entity_info(tileset)
-    }
-}
-
-impl LdtkEntity for LdtkSpriteSheetBundle {
-    fn bundle_entity(
-        entity_instance: &EntityInstance,
-        _: &LayerInstance,
-        tileset: Option<&Handle<Image>>,
-        tileset_definition: Option<&TilesetDefinition>,
-        _: &AssetServer,
-        texture_atlases: &mut Assets<TextureAtlasLayout>,
-    ) -> Self {
-        utils::sprite_sheet_bundle_from_entity_info(
-            entity_instance,
-            tileset,
-            tileset_definition,
-            texture_atlases,
-            true,
-        )
     }
 }
